Rename service-mvt-wfs to service-wfs-client
